requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-heroku==0.3.1
django-rest==0.8.7
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
whitenoise==6.9.0
